build(next): change release script
- includes separate `dev` and `beta` versions
- `dev` will work the same way as the main `dev` script that we have where we append the version with the timestamp
- `beta` will increment the version in the `/next/package.json` file
